🚸 Shorter stringified values for null-prototype (#5603)
**Description**

<!-- Please provide a short description and potentially linked issues
justifying the need for this PR -->

As specified on [MDN Web Doc @ Object # null-prototype
objects](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object#null-prototype_objects),
there are two main ways to build instances having null-prototype:

> However, you may create null-prototype objects using
Object.create(null) or the object initializer syntax with __proto__:
null [...]
>
> ```js
> const obj = Object.create(null);
> const obj2 = { __proto__: null };
> ````

Up-to-now, `fc.stringify` has been using the first notation with
`Object.assign` added to it to add keys within the instance. But as
`Object.create` + `Object.assign` tend to be longer than the
`___proto__:null`, we swicth to that shorter notation in that PR.

<!-- * Your PR is fixing a bug or regression? Check for existing issues
related to this bug and link them -->
<!-- * Your PR is adding a new feature? Make sure there is a related
issue or discussion attached to it -->

<!-- You can provide any additional context to help into understanding
what's this PR is attempting to solve: reproduction of a bug, code
snippets... -->

**Checklist** — _Don't delete this checklist and make sure you do the
following before opening the PR_

- [x] The name of my PR follows [gitmoji](https://gitmoji.dev/)
specification
- [x] My PR references one of several related issues (if any)
- [x] New features or breaking changes must come with an associated
Issue or Discussion
- [x] My PR does not add any new dependency without an associated Issue
or Discussion
- [x] My PR includes bumps details, please run `yarn bump` and flag the
impacts properly
- [x] My PR adds relevant tests and they would have failed without my PR
(when applicable)

<!-- More about contributing at
https://github.com/dubzzz/fast-check/blob/main/CONTRIBUTING.md -->

**Advanced**

<!-- How to fill the advanced section is detailed below! -->

- [x] Category: ✨ Introduce new features
- [x] Impacts: Lighter display for prototype-less instances

**⚠️ Minor breaking change**: As it may slighty impact final users
(exposed API), we will wait a major version to release it (major being
close, let's wait for it). It seems to be compatible with any version of
node, at least from 0.10.48 so the `eval(fc.stringify(value))` would
still work.

<!-- [Category] Please use one of the categories below, it will help us
into better understanding the urgency of the PR -->
<!-- * ✨ Introduce new features -->
<!-- * 📝 Add or update documentation -->
<!-- * ✅ Add or update tests -->
<!-- * 🐛 Fix a bug -->
<!-- * 🏷️ Add or update types -->
<!-- * ⚡️ Improve performance -->
<!-- * _Other(s):_ ... -->

<!-- [Impacts] Please provide a comma separated list of the potential
impacts that might be introduced by this change -->
<!-- * Generated values: Can your change impact any of the existing
generators in terms of generated values, if so which ones? when? -->
<!-- * Shrink values: Can your change impact any of the existing
generators in terms of shrink values, if so which ones? when? -->
<!-- * Performance: Can it require some typings changes on user side?
Please give more details -->
<!-- * Typings: Is there a potential performance impact? In which cases?
-->