🏷️ Better typings for `constant` (#5605)
**Description**

<!-- Please provide a short description and potentially linked issues
hustifying the need for this PR -->

As recommended by https://github.com/dubzzz/fast-check/issues/4047,
let's leverage const generics to make usage of `constant` even simpler
for TS users.

Not really a breaking change on its own but given we play with types and
change the way we declare some, let's be safe and pack this change with
other "next major" devs.

Requires at least TypeScript 5.0:
https://www.typescriptlang.org/docs/handbook/release-notes/typescript-5-0.html#const-type-parameters.

<!-- * Your PR is fixing a bug or regression? Check for existing issues
related to this bug and link them -->
<!-- * Your PR is adding a new feature? Make sure there is a related
issue or discussion attached to it -->

<!-- You can provide any additional context to help into understanding
what's this PR is attempting to solve: reproduction of a bug, code
snippets... -->

**Checklist** — _Don't delete this checklist and make sure you do the
following before opening the PR_

- [x] The name of my PR follows [gitmoji](https://gitmoji.dev/)
specification
- [x] My PR references one of several related issues (if any)
- [x] New features or breaking changes must come with an associated
Issue or Discussion
- [x] My PR does not add any new dependency without an associated Issue
or Discussion
- [x] My PR includes bumps details, please run `yarn bump` and flag the
impacts properly
- [x] My PR adds relevant tests and they would have failed without my PR
(when applicable)

<!-- More about contributing at
https://github.com/dubzzz/fast-check/blob/main/CONTRIBUTING.md -->

**Advanced**

<!-- How to fill the advanced section is detailed below! -->

- [x] Category: 🏷️ Add or update types
- [x] Impacts: Typings

<!-- [Category] Please use one of the categories below, it will help us
into better understanding the urgency of the PR --> <!-- * ✨ Introduce
new features -->
<!-- * 📝 Add or update documentation -->
<!-- * ✅ Add or update tests -->
<!-- * 🐛 Fix a bug -->
<!-- * 🏷️ Add or update types -->
<!-- * ⚡️ Improve performance -->
<!-- * _Other(s):_ ... -->

<!-- [Impacts] Please provide a comma separated list of the potential
impacts that might be introduced by this change --> <!-- * Generated
values: Can your change impact any of the existing generators in terms
of generated values, if so which ones? when? --> <!-- * Shrink values:
Can your change impact any of the existing generators in terms of shrink
values, if so which ones? when? --> <!-- * Performance: Can it require
some typings changes on user side? Please give more details -->
<!-- * Typings: Is there a potential performance impact? In which cases?
-->