⚡️ Faster initialization of globals by dropping `typeof` checks (#5617)
**Description**

<!-- Please provide a short description and potentially linked issues
justifying the need for this PR -->

**⚠️ Minor breaking change**

In the past we used to checks for `typeof Type` when initializing our
globals (the globals protecting fast-check from poisoning). It used to
be done because some of them only appeared starting at Node >8. With our
minimal target version being bumped to at least 10.5.0 we can safely
stop having these checks and save some precious time during tests (for
our users).

This change forces us to move minimal requirements to at least 10.5.0.

<!-- * Your PR is fixing a bug or regression? Check for existing issues
related to this bug and link them -->
<!-- * Your PR is adding a new feature? Make sure there is a related
issue or discussion attached to it -->

<!-- You can provide any additional context to help into understanding
what's this PR is attempting to solve: reproduction of a bug, code
snippets... -->

**Checklist** — _Don't delete this checklist and make sure you do the
following before opening the PR_

- [x] The name of my PR follows [gitmoji](https://gitmoji.dev/)
specification
- [x] My PR references one of several related issues (if any)
- [x] New features or breaking changes must come with an associated
Issue or Discussion
- [x] My PR does not add any new dependency without an associated Issue
or Discussion
- [x] My PR includes bumps details, please run `yarn bump` and flag the
impacts properly
- [x] My PR adds relevant tests and they would have failed without my PR
(when applicable)

<!-- More about contributing at
https://github.com/dubzzz/fast-check/blob/main/CONTRIBUTING.md -->

**Advanced**

<!-- How to fill the advanced section is detailed below! -->

- [x] Category: ⚡️ Improve performance
- [x] Impacts: We have to run against Node 10.5.0+

<!-- [Category] Please use one of the categories below, it will help us
into better understanding the urgency of the PR --> <!-- * ✨ Introduce
new features -->
<!-- * 📝 Add or update documentation -->
<!-- * ✅ Add or update tests -->
<!-- * 🐛 Fix a bug -->
<!-- * 🏷️ Add or update types -->
<!-- * ⚡️ Improve performance -->
<!-- * _Other(s):_ ... -->