💥 Extend `uuid` to build any know version (#5633)
**Description**

<!-- Please provide a short description and potentially linked issues
justifying the need for this PR -->

Officially UUID are supporting 15 versions: from 1 (inclusive) to 15
(inclusive). Among these versions only the first 8 ones are officially
provided with an implementation (or custom for the 8th one), the 7
others are for the moment not implemented.

As we want `uuid` to possibly generate any valid uuid we increased its
default to generate any uuid with versions 1 to 8 (it used to be 1 to
5).

This change is a breaking change as we might generate uuids for versions
6, 7 or 8 while it was not possible in the past except when requested
explicitly.

<!-- * Your PR is fixing a bug or regression? Check for existing issues
related to this bug and link them -->
<!-- * Your PR is adding a new feature? Make sure there is a related
issue or discussion attached to it -->

<!-- You can provide any additional context to help into understanding
what's this PR is attempting to solve: reproduction of a bug, code
snippets... -->

**Checklist** — _Don't delete this checklist and make sure you do the
following before opening the PR_

- [x] The name of my PR follows [gitmoji](https://gitmoji.dev/)
specification
- [x] My PR references one of several related issues (if any)
- [x] New features or breaking changes must come with an associated
Issue or Discussion
- [x] My PR does not add any new dependency without an associated Issue
or Discussion
- [x] My PR includes bumps details, please run `yarn bump` and flag the
impacts properly
- [x] My PR adds relevant tests and they would have failed without my PR
(when applicable)

<!-- More about contributing at
https://github.com/dubzzz/fast-check/blob/main/CONTRIBUTING.md -->

**Advanced**

<!-- How to fill the advanced section is detailed below! -->

- [x] Category: 💥 Breaking change
- [x] Impacts: Generated values can generate even more values than what
they used to