💥 Drop deprecated `ascii*` (#5636)
**Description**

<!-- Please provide a short description and potentially linked issues
justifying the need for this PR -->

The arbitraries `ascii` and `asciiString` have been marked for
deprecation since version 3.22.0.

They are now being fully dropped.

If you still want to rely on something behaving this way we recommend
you to use the following constructs:

```js
// for asciiString
string({ unit: 'binary-ascii' })

// for ascii
string({ unit: 'binary-ascii', minLength: 1, maxLength: 1 })
```

<!-- * Your PR is fixing a bug or regression? Check for existing issues
related to this bug and link them -->
<!-- * Your PR is adding a new feature? Make sure there is a related
issue or discussion attached to it -->

<!-- You can provide any additional context to help into understanding
what's this PR is attempting to solve: reproduction of a bug, code
snippets... -->

**Checklist** — _Don't delete this checklist and make sure you do the
following before opening the PR_

- [x] The name of my PR follows [gitmoji](https://gitmoji.dev/)
specification
- [x] My PR references one of several related issues (if any)
- [x] New features or breaking changes must come with an associated
Issue or Discussion
- [x] My PR does not add any new dependency without an associated Issue
or Discussion
- [x] My PR includes bumps details, please run `yarn bump` and flag the
impacts properly
- [x] My PR adds relevant tests and they would have failed without my PR
(when applicable)

<!-- More about contributing at
https://github.com/dubzzz/fast-check/blob/main/CONTRIBUTING.md -->

**Advanced**

<!-- How to fill the advanced section is detailed below! -->

- [x] Category: 💥 Breaking change
- [x] Impacts: Dropped arbitraries