👷 Move build chain to ESM (#5670)
**Description**

<!-- Please provide a short description and potentially linked issues
justifying the need for this PR -->

The build chain of fast-check has been CommonJS-based since day 1. With
ESM moving forward in the ecosystem, it's time to move ourselves to the
new standard and adapt our build chains to ESM.

Unfortunately it may have some subtle impacts on our users as our
package will not be a CJS one offering a ESM fallback anymore. I will
rather be the opposite: an ESM package with a fallback to CJS. It
implies that we moved ESM related files closer to the root of the
package (we could have kept them in esm/) and moved the CJS ones further
in the file structure (we had to move them).

Another subtle impact is that it would impose our users to run at least
Node ≥12.17.0.

As such we consider it as a breaking change. On its own it should not be
a huge problem for most of the users but given it changes our minimal
requirement let's put it into the next major release.

Superseed #4592

<!-- * Your PR is fixing a bug or regression? Check for existing issues
related to this bug and link them -->
<!-- * Your PR is adding a new feature? Make sure there is a related
issue or discussion attached to it -->

<!-- You can provide any additional context to help into understanding
what's this PR is attempting to solve: reproduction of a bug, code
snippets... -->

**Checklist** — _Don't delete this checklist and make sure you do the
following before opening the PR_

- [x] The name of my PR follows [gitmoji](https://gitmoji.dev/)
specification
- [x] My PR references one of several related issues (if any)
- [x] New features or breaking changes must come with an associated
Issue or Discussion
- [x] My PR does not add any new dependency without an associated Issue
or Discussion
- [x] My PR includes bumps details, please run `yarn bump` and flag the
impacts properly
- [x] My PR adds relevant tests and they would have failed without my PR
(when applicable)

<!-- More about contributing at
https://github.com/dubzzz/fast-check/blob/main/CONTRIBUTING.md -->

**Advanced**

<!-- How to fill the advanced section is detailed below! -->

- [x] Category: 👷 Configuration of the package
- [x] Impacts: Requires users to rely on more recent versions of Node,
at least 12.17

<!-- [Category] Please use one of the categories below, it will help us
into better understanding the urgency of the PR -->
<!-- * ✨ Introduce new features -->
<!-- * 📝 Add or update documentation -->
<!-- * ✅ Add or update tests -->
<!-- * 🐛 Fix a bug -->
<!-- * 🏷️ Add or update types -->
<!-- * ⚡️ Improve performance -->
<!-- * _Other(s):_ ... -->

<!-- [Impacts] Please provide a comma separated list of the potential
impacts that might be introduced by this change -->
<!-- * Generated values: Can your change impact any of the existing
generators in terms of generated values, if so which ones? when? -->
<!-- * Shrink values: Can your change impact any of the existing
generators in terms of shrink values, if so which ones? when? -->
<!-- * Performance: Can it require some typings changes on user side?
Please give more details -->
<!-- * Typings: Is there a potential performance impact? In which cases?
-->